Reorganized file structure for /investments product to keep a more consistent design pattern. Bumped version.
