+++ README.md
@@ -25,7 +25,7 @@ Add to your dependencies in `mix.exs`. The hex specification is required.
 
 ```elixir
 def deps do
-  [{:plaid, "~> 1.6", hex: :plaid_elixir}]
+  [{:plaid, "~> 1.7", hex: :plaid_elixir}]
 end
 ```
